ReCaptcha done

+++ Ogma3/Pages/Stories/Delete.cshtml.cs
@@ -54,6 +54,7 @@ namespace Ogma3.Pages.Stories
             Story = await _context.Stories
                 .Include(s => s.Author)
                 .Include(s => s.Chapters)
+                    .ThenInclude(c => c.CommentsThread)
                 .Include(s => s.VotesPool)
                     .ThenInclude(vp => vp.Votes)
                 .FirstOrDefaultAsync(s => s.Id == id && s.Author.IsLoggedIn(User));
@@ -66,13 +67,17 @@ namespace Ogma3.Pages.Stories
 
             // Remove story
             _context.Stories.Remove(Story);
-            // // Remove votes
-            // _context.Votes.RemoveRange(Story.VotesPool.Votes);
-            // // Remove votes pool
-            // _context.VotePools.Remove(Story.VotesPool);
+            
+            // Remove votes
+            _context.Votes.RemoveRange(Story.VotesPool.Votes);
+            // Remove votes pool
+            _context.VotePools.Remove(Story.VotesPool);
+            
+            // Remove CommentThreads
+            _context.CommentThreads.RemoveRange(Story.Chapters.Select(c => c.CommentsThread).ToList());
 
             // Delete cover
-            if (Story.CoverId != null)
+            if (Story.CoverId != null && Story.Cover != null) 
                 await _b2Client.Files.Delete(Story.CoverId, Story.Cover.Replace(_config["cdn"], ""));
 
             // Save
